+++ src/model_runners/model_runner.py
@@ -1,8 +1,6 @@
 from abc import ABC, abstractmethod
 from typing import Dict, Tuple, List
 
-from ray.data import Dataset
-
 
 class ModelRunner(ABC):
     """
@@ -16,19 +14,16 @@ class ModelRunner(ABC):
     def predict(self, prompt: str) -> Tuple[str, float]:
         """
         Runs the model on the given data. Output and log-probability will both be written
-        to the `prediction_column_name` column as dictionary. Ray can handle nested
-        dicts in the Dataset.
+        to the `prediction_column_name` column as dictionary. It returns a tuple of:
 
-        The nested dict will have the structure:
         ```
-        {
+        (
             "output": str
             "input_log_probability": float
-        }
+        )
         ```
-        The output values are extracted using the `output` parameter provided in the analysis configuration file.
-        Similarly, probability values are extracted from the model response using the `probability` parameter in the
-        analysis configuration file.
+        The output values are extracted using the `output` parameter provided in the model settings.
+        Similarly, probability values are extracted from the model response using the `probability` parameter.
 
         :param data: the dataset where each instance is a row in the dataset.
         :param prompt: the prompt
