+++ src/model_runners/util.py
@@ -1,0 +1,119 @@
+"""
+Utilities for model runners.
+"""
+import atexit
+import logging
+import time
+
+from abc import ABC, abstractmethod
+
+import sagemaker
+from sagemaker.utils import unique_name_from_base
+from sagemaker.jumpstart.model import JumpStartModel as SDKJumpStartModel
+from sagemaker.predictor import retrieve_default
+from sagemaker.base_predictor import PredictorBase
+from typing import Optional
+from dataclasses import dataclass
+
+logger = logging.getLogger(__name__)
+
+@dataclass
+class SagemakerEndpointUsageState:
+    """Endpoint usage state config use by MangedEndpoint.
+
+    :param start_time: start time of managing this shadow endpoint.
+    :param num_calls: Number of calls of waiting for this endpoint spinning up.
+    :param in_service: whether the endpoint is in service or not.
+    """
+
+    start_time: float
+    num_calls: int
+    in_service: bool
+
+class SagemakerModelInvocationParameters(ABC):
+    """
+    Parameter(s) to be used for model invocation.
+    """
+    endpoint_name: str
+    model_id: str
+    model_name: str
+    predictor: PredictorBase
+
+def wait_for_sagemaker_endpoint(
+    sagemaker_session: sagemaker.session.Session,
+    endpoint_name: str,
+    endpoint_usage_state: SagemakerEndpointUsageState,
+) -> SagemakerEndpointUsageState:
+    """
+    Wait for SageMaker shadow endpoint deployment to complete.
+    :param sagemaker_session: SageMaker session to be reused.
+    :param endpoint_name: SageMaker shadow endpoint name.
+    :param endpoint_usage_state: Endpoint usage state config use by MangedEndpoint.
+    :return EndpointUsageState: Updated endpoint usage state config use by MangedEndpoint.
+    """
+    from inspect import cleandoc
+
+    if not endpoint_usage_state.in_service:  # pragma: no branch
+        # When should_wait is true we always make sure the endpoint is ready.
+        # Otherwise, we only wait if we created a new shadow endpoint (in this case
+        # model_name is set).
+        wait_time_start = time.time()
+        logger.info(
+            cleandoc(
+                """Checking endpoint status:
+        Legend:
+        (OutOfService: x, Creating: -, Updating: -, InService: !, RollingBack: <, Deleting: o, Failed: *)"""
+            )
+        )
+        # Do initial check to see if endpoint is already up and running before wait polling
+        # See https://tiny.amazon.com/6uiwux44/issuamazissuRAI4 for why this is needed
+        desc = sagemaker_session.sagemaker_client.describe_endpoint(EndpointName=endpoint_name)
+        if not desc or "EndpointStatus" not in desc or desc["EndpointStatus"] != "InService":
+            # NOTE: Endpoint creation usually takes 5 to 12 minutes, here poll status every minute.
+            sagemaker_session.wait_for_endpoint(endpoint_name, poll=60)
+        wait_time = time.time() - wait_time_start
+        logger.info("Endpoint is in service after %.0f seconds", wait_time)
+    endpoint_usage_state.in_service = True
+    endpoint_usage_state.start_time = time.time()
+    endpoint_usage_state.num_calls += 1
+    return endpoint_usage_state
+
+
+def gen_shadow_endpoint_unique_name(prefix: str, model_name: str, unique_len: int = 15) -> str:
+    """
+    Generate shadow endpoint unique name.
+    :param prefix: endpoint name prefix
+    :param model_name: model name
+    :param unique_len: unique characters length
+    :return: endpoint name
+    """
+    rem_len = ENDPOINT_NAME_MAX_LEN - unique_len
+    if len(prefix) + len(model_name) > rem_len:
+        prefix_len = (rem_len // 2) - 1
+        prefix = prefix[:prefix_len]
+        model_name_len = rem_len - len(prefix)
+        model_name = model_name[:model_name_len]
+    endpoint_name = unique_name_from_base(prefix + "-" + model_name, ENDPOINT_NAME_MAX_LEN)
+    return endpoint_name
+
+def create_content_composer(template: str) -> Optional[ContentComposer]:
+    composer: Optional[ContentComposer] = None
+    vanilla_template = VanillaTemplate(template)
+
+    if identifiers := vanilla_template.get_unique_identifiers():
+        if SinglePromptContentComposer.KEYWORD in identifiers:
+            composer = SinglePromptContentComposer(vanilla_template)
+        elif MultiPromptsContentComposer.KEYWORD in identifiers:
+            composer = MultiPromptsContentComposer(vanilla_template)
+        else:
+            logger.error(f"Found placeholders {identifiers} in template '{template}'.")
+    else:
+        logger.error(f"Could not find any identifier in template '{template}'.")
+
+    if composer is None:
+        raise UserError(
+            "Invalid content_template. View job logs for details. "
+            "The template must contain the placeholder $prompts for an array of prompts, "
+            "or the placeholder $prompt for single prompt at a time."
+        )
+    return composer
